+++ src/main/java/com/dh/hospital/repository/HospitalRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dh.hospital.entity.Hospital;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
 
 
 }
