+++ src/main/java/com/dh/hospital/repository/NotaVisitaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dh.hospital.entity.NotaVisita;
 
-public interface NotaVisitaRepository extends JpaRepository<NotaVisita, Integer> {
+public interface NotaVisitaRepository extends JpaRepository<NotaVisita, Long> {
 
 }
